+++ src/Resources/app/storefront/src/main.js
@@ -1,5 +1,11 @@
 import CartFlyoutPlugin from "./plugin/cart-flyout/cart-flyout.plugin";
+import UspRotatorPlugin from "./plugin/usp-rotator/usp-rotator.plugin";
 
 const PluginManager = window.PluginManager;
 
 PluginManager.register("CartFlyoutPlugin", CartFlyoutPlugin, ".offcanvas-cart");
+PluginManager.register(
+    "UspRotator",
+    UspRotatorPlugin,
+    '[data-plugin="UspRotator"]'
+);